Move Appearance font coverage into the public home suite
